refactor: move most api logic into App
